% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_classes.R
\name{are_equivalent}
\alias{are_equivalent}
\title{Check if two models are equivalent}
\usage{
are_equivalent(x, model1, model2)
}
\arguments{
\item{x}{An \code{equiv_classes} object}

\item{model1}{Name of first model}

\item{model2}{Name of second model}
}
\value{
Logical
}
\description{
Check if two models are equivalent
}
