% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{can_simulate}
\alias{can_simulate}
\title{Check if model_spec can simulate}
\usage{
can_simulate(spec)
}
\arguments{
\item{spec}{A \code{model_spec} object}
}
\value{
Logical
}
\description{
Check if model_spec can simulate
}
