% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_classes.R
\name{class_members}
\alias{class_members}
\title{Get members of an equivalence class}
\usage{
class_members(x, class_id)
}
\arguments{
\item{x}{An \code{equiv_classes} object}

\item{class_id}{Class index (integer)}
}
\value{
Character vector of model names
}
\description{
Get members of an equivalence class
}
