% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{coef.fitted_model_spec}
\alias{coef.fitted_model_spec}
\title{Extract coefficients from fitted model_spec}
\usage{
\method{coef}{fitted_model_spec}(object, ...)
}
\arguments{
\item{object}{A \code{fitted_model_spec} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Named numeric vector of parameter estimates
}
\description{
Extract coefficients from fitted model_spec
}
