% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degen-package.R
\docType{package}
\name{degen-package}
\alias{degen}
\alias{degen-package}
\title{degen: Detect Observational Equivalence and Non-Identifiability in Models}
\description{
Tools for detecting whether statistical model specifications are observationally equivalent, diagnosing identifiability problems, and characterizing equivalence classes among competing formulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcol33/degen}
  \item Report bugs at \url{https://github.com/gcol33/degen/issues}
}

}
\author{
\strong{Maintainer}: Gilles Colling \email{gilles.colling@uclouvain.be}

}
\keyword{internal}
