% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{info_condition}
\alias{info_condition}
\title{Get condition number of Fisher information}
\usage{
info_condition(x)
}
\arguments{
\item{x}{A \code{fisher_info} object}
}
\value{
Numeric scalar
}
\description{
Get condition number of Fisher information
}
