% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{info_eigenvalues}
\alias{info_eigenvalues}
\title{Extract eigenvalues from Fisher information}
\usage{
info_eigenvalues(x)
}
\arguments{
\item{x}{A \code{fisher_info} object}
}
\value{
Numeric vector of eigenvalues (sorted decreasing)
}
\description{
Extract eigenvalues from Fisher information
}
