% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{info_rank}
\alias{info_rank}
\title{Get numerical rank of Fisher information}
\usage{
info_rank(x)
}
\arguments{
\item{x}{A \code{fisher_info} object}
}
\value{
Integer
}
\description{
Get numerical rank of Fisher information
}
