% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_classes.R
\name{is_equiv_classes}
\alias{is_equiv_classes}
\title{Test if object is an equiv_classes}
\usage{
is_equiv_classes(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if object is an equiv_classes
}
