% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_pair.R
\name{is_equivalence_pair}
\alias{is_equivalence_pair}
\title{Test if object is an equivalence_pair}
\usage{
is_equivalence_pair(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether \code{x} is an \code{equivalence_pair} object
}
\description{
Test if object is an equivalence_pair
}
\examples{
exp_spec <- model_spec(
  loglik_fn = function(y, lambda) sum(dexp(y, rate = lambda, log = TRUE)),
  par_names = "lambda"
)
pair <- equivalence_pair(exp_spec, exp_spec)
is_equivalence_pair(pair)  # TRUE
is_equivalence_pair(list())
is_equivalence_pair(list())  # FALSE
}
