% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{is_fisher_info}
\alias{is_fisher_info}
\title{Test if object is a fisher_info}
\usage{
is_fisher_info(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if object is a fisher_info
}
