% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{is_identifiability_result}
\alias{is_identifiability_result}
\title{Test if object is an identifiability_result}
\usage{
is_identifiability_result(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if object is an identifiability_result
}
