% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec-methods.R
\name{is_model_spec}
\alias{is_model_spec}
\title{Test if object is a model_spec}
\usage{
is_model_spec(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether \code{x} is a \code{model_spec} object
}
\description{
Test if object is a model_spec
}
\examples{
spec <- model_spec(
  loglik_fn = function(y, lambda) sum(dexp(y, rate = lambda, log = TRUE)),
  par_names = "lambda"
)
is_model_spec(spec)  # TRUE
is_model_spec(list())  # FALSE
}
