% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_comparison.R
\name{is_surface_comparison}
\alias{is_surface_comparison}
\title{Test if object is a surface_comparison}
\usage{
is_surface_comparison(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if object is a surface_comparison
}
