% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/model_spec-methods.R
\name{loglik}
\alias{loglik}
\alias{loglik.model_spec}
\title{Evaluate log-likelihood}
\usage{
loglik(object, y, par, ...)

\method{loglik}{model_spec}(object, y, par, ...)
}
\arguments{
\item{object}{A model specification object}

\item{y}{Numeric vector of observed data}

\item{par}{Named numeric vector of parameter values}

\item{...}{Additional arguments passed to methods}
}
\value{
Numeric scalar log-likelihood value
}
\description{
Evaluate the log-likelihood function for a model specification at given
parameter values.
}
\examples{
spec <- model_spec(
  loglik_fn = function(y, lambda) sum(dexp(y, rate = lambda, log = TRUE)),
  par_names = "lambda",
  par_bounds = list(lambda = c(1e-6, Inf))
)
y <- rexp(100, rate = 2)
loglik(spec, y, par = c(lambda = 2))
}
