% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_beta}
\alias{model_spec_beta}
\title{Create model_spec for Beta distribution}
\usage{
model_spec_beta(name = "Beta")
}
\arguments{
\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Beta distribution
}
\examples{
spec <- model_spec_beta()
y <- rbeta(100, shape1 = 2, shape2 = 5)
loglik(spec, y, c(shape1 = 2, shape2 = 5))
}
