% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_binomial}
\alias{model_spec_binomial}
\title{Create model_spec for Binomial distribution}
\usage{
model_spec_binomial(size, name = "Binomial")
}
\arguments{
\item{size}{Number of trials (known)}

\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Binomial distribution
}
\examples{
spec <- model_spec_binomial(size = 10)
y <- rbinom(100, size = 10, prob = 0.3)
loglik(spec, y, c(prob = 0.3))
}
