% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{model_spec_from_fit}
\alias{model_spec_from_fit}
\title{Extract model_spec from fitted R objects}
\usage{
model_spec_from_fit(object, ...)
}
\arguments{
\item{object}{A fitted model object (lm, glm, nls, etc.)}

\item{...}{Additional arguments (currently unused)}
}
\value{
A \code{model_spec} object
}
\description{
Create a model_spec from common fitted model objects.
}
\examples{
# From linear model
fit <- lm(mpg ~ wt, data = mtcars)
spec <- model_spec_from_fit(fit)
print(spec)
}
