% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_negbinom}
\alias{model_spec_negbinom}
\title{Create model_spec for Negative Binomial distribution}
\usage{
model_spec_negbinom(par = c("size_prob", "size_mu"), name = "NegBinomial")
}
\arguments{
\item{par}{Parameterization: "size_prob" or "size_mu" (mean parameterization)}

\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Negative Binomial distribution
}
\examples{
spec <- model_spec_negbinom()
y <- rnbinom(100, size = 5, prob = 0.3)
loglik(spec, y, c(size = 5, prob = 0.3))
}
