% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_poisson}
\alias{model_spec_poisson}
\title{Create model_spec for Poisson distribution}
\usage{
model_spec_poisson(name = "Poisson")
}
\arguments{
\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Poisson distribution
}
\examples{
spec <- model_spec_poisson()
y <- rpois(100, lambda = 5)
loglik(spec, y, c(lambda = 5))
}
