% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_uniform}
\alias{model_spec_uniform}
\title{Create model_spec for Uniform distribution}
\usage{
model_spec_uniform(name = "Uniform")
}
\arguments{
\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Uniform distribution
}
\examples{
spec <- model_spec_uniform()
y <- runif(100, min = 2, max = 8)
loglik(spec, y, c(min = 2, max = 8))
}
