% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec_helpers.R
\name{model_spec_weibull}
\alias{model_spec_weibull}
\title{Create model_spec for Weibull distribution}
\usage{
model_spec_weibull(name = "Weibull")
}
\arguments{
\item{name}{Optional model name}
}
\value{
A \code{model_spec} object
}
\description{
Create model_spec for Weibull distribution
}
\examples{
spec <- model_spec_weibull()
y <- rweibull(100, shape = 2, scale = 1)
loglik(spec, y, c(shape = 2, scale = 1))
}
