% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_classes.R
\name{n_classes}
\alias{n_classes}
\title{Get number of equivalence classes}
\usage{
n_classes(x)
}
\arguments{
\item{x}{An \code{equiv_classes} object}
}
\value{
Integer
}
\description{
Get number of equivalence classes
}
