% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/model_spec-methods.R
\name{n_par}
\alias{n_par}
\alias{n_par.model_spec}
\title{Get number of parameters}
\usage{
n_par(x, ...)

\method{n_par}{model_spec}(x, ...)
}
\arguments{
\item{x}{An object containing parameters}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer number of parameters
}
\description{
Get number of parameters
}
