% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_pair.R
\name{pair_accessors}
\alias{pair_accessors}
\alias{spec_a}
\alias{spec_b}
\title{Extract model specification from pair}
\usage{
spec_a(pair)

spec_b(pair)
}
\arguments{
\item{pair}{An \code{equivalence_pair} object}
}
\value{
A \code{model_spec} object
}
\description{
Extract model specification from pair
}
