% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/model_spec-methods.R
\name{par_bounds}
\alias{par_bounds}
\alias{par_bounds.model_spec}
\title{Get parameter bounds}
\usage{
par_bounds(x, ...)

\method{par_bounds}{model_spec}(x, ...)
}
\arguments{
\item{x}{An object containing parameter bounds}

\item{...}{Additional arguments passed to methods}
}
\value{
Named list of parameter bounds
}
\description{
Get parameter bounds
}
