% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_pair.R
\name{par_dims}
\alias{par_dims}
\title{Get parameter dimensions for both models}
\usage{
par_dims(pair)
}
\arguments{
\item{pair}{An \code{equivalence_pair} object}
}
\value{
Named integer vector with elements \code{a} and \code{b}
}
\description{
Get parameter dimensions for both models
}
\examples{
exp_spec <- model_spec(
  loglik_fn = function(y, lambda) sum(dexp(y, rate = lambda, log = TRUE)),
  par_names = "lambda"
)
norm_spec <- model_spec(
  loglik_fn = function(y, mu, sigma) sum(dnorm(y, mu, sigma, log = TRUE)),
  par_names = c("mu", "sigma")
)
pair <- equivalence_pair(exp_spec, norm_spec)
par_dims(pair)  # c(a = 1, b = 2)
}
