% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/model_spec-methods.R
\name{par_names}
\alias{par_names}
\alias{par_names.model_spec}
\title{Get parameter names}
\usage{
par_names(x, ...)

\method{par_names}{model_spec}(x, ...)
}
\arguments{
\item{x}{An object containing parameter information}

\item{...}{Additional arguments passed to methods}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names
}
