% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{profile_likelihood}
\alias{profile_likelihood}
\title{Compute profile likelihood for one parameter}
\usage{
profile_likelihood(spec, y, par, which_par, n_points = 20)
}
\arguments{
\item{spec}{A model_spec object}

\item{y}{Data}

\item{par}{Starting parameter values}

\item{which_par}{Name of parameter to profile}

\item{n_points}{Number of points in profile}
}
\value{
Data frame with profile likelihood values
}
\description{
Compute profile likelihood for one parameter
}
