% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_pair.R
\name{same_par_count}
\alias{same_par_count}
\title{Check if models have same parameter count}
\usage{
same_par_count(pair)
}
\arguments{
\item{pair}{An \code{equivalence_pair} object}
}
\value{
Logical
}
\description{
Check if models have same parameter count
}
