% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stop parallel cluster}
\usage{
stop_cluster(cl)
}
\arguments{
\item{cl}{Cluster object from \code{setup_cluster()}}
}
\value{
NULL invisibly
}
\description{
Stop parallel cluster
}
