% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{vcov.fitted_model_spec}
\alias{vcov.fitted_model_spec}
\title{Extract variance-covariance matrix}
\usage{
\method{vcov}{fitted_model_spec}(object, ...)
}
\arguments{
\item{object}{A \code{fitted_model_spec} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Variance-covariance matrix
}
\description{
Compute variance-covariance matrix from the inverse Hessian.
}
